move action 'types' variables to own folder
