+++ src/stateManagement/reducers/flagsReducer.js
@@ -1,5 +1,18 @@
-import { userConstants } from '../actions/registrationTypes';
-import { LISTINGS_REQUEST, LISTINGS_SUCCESS, LISTINGS_FAILURE, ADD_LISTING_REQUESTED, ADD_LISTING_SUCCESSFUL, ADD_LISTING_FAILURE, EDIT_LISTING_REQUESTED, EDIT_LISTING_SUCCESSFUL, EDIT_LISTING_FAILURE, DELETE_LISTING_REQUESTED, DELETE_LISTING_SUCCESSFUL, DELETE_LISTING_FAILURE } from '../actions/types';
+import { userConstants } from '../types/registrationTypes';
+import {
+  LISTINGS_REQUEST,
+  LISTINGS_SUCCESS,
+  LISTINGS_FAILURE,
+  ADD_LISTING_REQUESTED,
+  ADD_LISTING_SUCCESSFUL,
+  ADD_LISTING_FAILURE,
+  EDIT_LISTING_REQUESTED,
+  EDIT_LISTING_SUCCESSFUL,
+  EDIT_LISTING_FAILURE,
+  DELETE_LISTING_REQUESTED,
+  DELETE_LISTING_SUCCESSFUL,
+  DELETE_LISTING_FAILURE
+} from '../types/types';
 
 const initalFlags = {
   loginRequested: false,
