+++ src/stateManagement/reducers/listingsReducer.js
@@ -1,4 +1,4 @@
-import { LISTINGS_SUCCESS, LISTINGS_FAILURE, ADD_LISTING_SUCCESSFUL, EDIT_LISTING_SUCCESSFUL, DELETE_LISTING_SUCCESSFUL } from "../actions/types";
+import { LISTINGS_SUCCESS, LISTINGS_FAILURE, ADD_LISTING_SUCCESSFUL, EDIT_LISTING_SUCCESSFUL, DELETE_LISTING_SUCCESSFUL } from "../types/types";
 
 export default (state = {}, action) => {
   switch (action.type) {
