+++ src/stateManagement/types/registrationTypes.js
